requirements.txt
absl-py==2.1.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anykeystore==0.2
apex==0.9.10.dev0
appdirs==1.4.4
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi 
charset-normalizer==3.4.2
click==8.2.1
colorful==0.5.7
comm 
contourpy==1.3.2
crcmod==1.7
cryptacular==1.6.2
cycler==0.12.1
debugpy 
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
edt==3.0.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.14.0
fonttools==4.59.0
fsspec==2025.7.0
gdown==5.2.0
greenlet==3.0.3
grpcio==1.65.2
hf-xet==1.1.5
huggingface-hub==0.33.4
humanize==4.12.3
hupper==1.12.1
idna==3.10
imageio==2.37.0
importlib_metadata 
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel 
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
jupyter_client 
jupyter_core 
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
medim==0.1.2
model-index==0.1.11
monai==1.5.0
mpmath==1.3.0
msgpack==1.0.8
nest_asyncio 
networkx==3.4.2
nibabel==5.3.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.3.1
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
ordered-set==4.1.0
oss2==2.17.0
packaging 
pandas==2.3.1
parso==0.8.4
PasteDeploy==3.1.0
pbkdf2==1.3
pexpect==4.9.0
pillow 
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs 
pluggy==1.6.0
prefetch_generator==1.0.3
prettytable==3.16.0
prompt_toolkit==3.0.51
protobuf==4.25.4
psutil 
ptyprocess==0.7.0
pudb==2025.1
pure_eval==0.2.3
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynvim==0.5.0
pyparsing==3.2.3
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.4
PyYAML==6.0.2
pyzmq 
repoze.sendmail==4.4.1
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
scikit-image==0.25.2
scipy==1.15.3
shellingham==1.5.4
simpleitk==2.5.2
simsimd==6.5.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
stringzilla==3.12.5
surface-distance==0.1
sympy==1.13.1
tabulate==0.9.0
tbparse==0.0.9
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tifffile==2025.5.10
tomli==2.2.1
torch==2.6.0
torch-tb-profiler==0.4.3
torchaudio==2.6.0
torchio==0.20.19
torchvision==0.21.0
tornado 
tqdm==4.65.2
traitlets==5.14.3
transaction==5.0
translationstring==1.4
triton==3.2.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==1.26.20
urwid==3.0.2
urwid_readline==0.15.1
uv==0.8.0
velruse==1.1.1
venusian==3.1.1
wcwidth==0.2.13
WebOb==1.8.9
Werkzeug==3.0.3
wrapt==1.17.2
WTForms==3.2.1
wtforms-recaptcha==0.3.2
yacs==0.1.8
zipp 
zope.deprecation==5.1
zope.interface==7.2
zope.sqlalchemy==3.1
